cors + modificacion en links + dump + agregar numero partido y id partido en getResultados

+++ grails-app/utils/dto/response/ResultadoMesaDTO.groovy
@@ -4,7 +4,7 @@ package dto.response
  * Created by pablo on 07/10/17.
  */
 class ResultadoMesaDTO {
-    Long mesa
+    Integer mesa
     List<ResultadoPartidoMesaDTO> resultados
 
     public ResultadoMesaDTO(){
